test(radio): separate hook test suites into 1:1 dedicated test files
